+++ web/jobs/assets/BrowserView-es_kl5ql.js
@@ -1,4 +1,4 @@
-import{c as j,r,I as X,J as Z,j as e,i as w,t as ee,T as se,K as ie,C as $,A as x,N as ne,D as g}from"./index-CpKprEnV.js";import{i as ae,r as le,C as re,E as ce}from"./application-flow-Bfyr1zTm.js";import{P as B}from"./play-Clxw3iUv.js";import{M as D}from"./monitor-up-DA12hCSB.js";/**
+import{c as j,r,I as X,J as Z,j as e,i as w,t as ee,T as se,K as ie,C as $,A as x,N as ne,D as g}from"./index-BK3c0l6M.js";import{i as ae,r as le,C as re,E as ce}from"./application-flow-ofRzYjOl.js";import{P as B}from"./play-DsCGhNz9.js";import{M as D}from"./monitor-up-B6L--0ro.js";/**
  * @license lucide-react v0.500.0 - ISC
  *
  * This source code is licensed under the ISC license.