+++ web/jobs/assets/SettingsView-BJngCQe5.js
@@ -1,4 +1,4 @@
-import{c as $,r as n,j as e,P as X,a as Ce,K as G,d as We,O as Z,D as V,Q as z,m as Q,g as Je,C as de,M as Se,F as Ke,i as ke,U as Qe,V as ue,v as D,W as Oe,X as Xe,Y as ee,Z as qe,$ as Pe,a0 as Ze,t as $e,A as es,a1 as ss,a2 as is,a3 as we,a4 as ns,a5 as as,z as ts,k as Ee}from"./index-CpKprEnV.js";import{M as he,C as ls}from"./mail-CrlsSWeB.js";import{C as Le}from"./chevron-right-CVxl_yhm.js";import{M as os}from"./monitor-up-DA12hCSB.js";import{P as cs}from"./pencil-BW3zfdKH.js";/**
+import{c as $,r as n,j as e,P as X,a as Ce,K as G,d as We,O as Z,D as V,Q as z,m as Q,g as Je,C as de,M as Se,F as Ke,i as ke,U as Qe,V as ue,v as D,W as Oe,X as Xe,Y as ee,Z as qe,$ as Pe,a0 as Ze,t as $e,A as es,a1 as ss,a2 as is,a3 as we,a4 as ns,a5 as as,z as ts,k as Ee}from"./index-BK3c0l6M.js";import{M as he,C as ls}from"./mail-Dw4VxV1K.js";import{C as Le}from"./chevron-right-CzoJ7jOh.js";import{M as os}from"./monitor-up-B6L--0ro.js";import{P as cs}from"./pencil-CVlkRX_y.js";/**
  * @license lucide-react v0.500.0 - ISC
  *
  * This source code is licensed under the ISC license.