+++ web/jobs/assets/chevron-right-CzoJ7jOh.js
@@ -1,4 +1,4 @@
-import{c as o}from"./index-CpKprEnV.js";/**
+import{c as o}from"./index-BK3c0l6M.js";/**
  * @license lucide-react v0.500.0 - ISC
  *
  * This source code is licensed under the ISC license.