+++ web/jobs/assets/monitor-up-B6L--0ro.js
@@ -1,4 +1,4 @@
-import{c as t}from"./index-CpKprEnV.js";/**
+import{c as t}from"./index-BK3c0l6M.js";/**
  * @license lucide-react v0.500.0 - ISC
  *
  * This source code is licensed under the ISC license.